Added prototype wide (Unicode) string support.
git-svn-id: svn://localhost/public/MoSync/trunk@658 6ad7a17e-d552-4a1e-ace4-f958fe14b117

+++ runtimes/cpp/base/Syscall.cpp
@@ -69,7 +69,7 @@ namespace Base {
 #ifdef IX_FILE
 	struct FileList {
 		std::vector<std::string> files;
-		int pos;
+		size_t pos;
 	};
 	typedef std::map<int, FileList> FileListMap;
 	typedef FileListMap::iterator FileListItr;
